Moved BaseGlobalExporter to its own file
